+++ CHANGELOG.md
@@ -10,9 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   set to `True`, targets seek to avoid the given `match_value` instead of matching it.
 - Transfer learning regression benchmarks infrastructure for evaluating TL model
   performance on regression tasks
+- `Objective.to_botorch_posterior_transform` for use of affine transformations with
+  analytical acquisition functions
+
+### Changed
+- `DesirabilityObjective.weights` now returns the normalized target weights
 
 ### Fixed
 - Random seed not entering simulation when explicitly passed to `simulate_scenarios`
+- `DesirabilityObjective` now properly interplays with analytical acquisition functions
+
+### Removed
+- `AffinePostTransformation` class (since BoTorch provides equivalent functionality)
  
 ## [0.14.1] - 2025-10-01
 ### Added